Move s-u to PSR-4
