+++ tests/lib/Security/CSP/ContentSecurityPolicyManagerTest.php
