+++ tests/lib/Security/CSRF/CsrfTokenGeneratorTest.php
