+++ tests/lib/Security/CSRF/CsrfTokenManagerTest.php
