+++ tests/lib/Security/CertificateManagerTest.php
