+++ tests/lib/Security/CertificateTest.php
