+++ tests/lib/Security/CredentialsManagerTest.php
