+++ tests/lib/Security/CryptoTest.php
