+++ tests/lib/Security/HasherTest.php
