+++ tests/lib/Security/SecureRandomTest.php
