+++ tests/lib/Security/TrustedDomainHelperTest.php
