+++ tests/lib/Session/CryptoSessionDataTest.php
