+++ tests/lib/Session/CryptoWrappingTest.php
