+++ tests/lib/Share/MailNotificationsTest.php
