+++ tests/lib/Share/SearchResultSorterTest.php
@@ -21,7 +21,7 @@
 
 namespace Test\Share;
 
-class SearchResultSorter extends \Test\TestCase {
+class SearchResultSorterTest extends \Test\TestCase {
 	public function testSort() {
 		$search = 'lin';
 		$sorter = new \OC\Share\SearchResultSorter($search, 'foobar');
