+++ tests/lib/Share20/DefaultShareProviderTest.php
