+++ tests/lib/Share20/ManagerTest.php
