+++ tests/lib/Traits/EncryptionTrait.php
