+++ tests/lib/Traits/UserTrait.php
