+++ tests/lib/User/AvatarUserDummy.php
