+++ tests/lib/User/Dummy.php
