+++ tests/lib/Util/Group/Dummy.php
