+++ tests/lib/Util/User/Dummy.php
